pyproject.toml
[project]
name = "post-training-and-alignment"
version = "0.1.0"
description = "Post-training pipeline: SFT, DPO, RLHF, GRPO on small language models"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.11.0",
    "transformers>=4.50.0",
    "peft>=0.15.0",
    "trl>=0.18.0",
    "datasets>=3.5.0",
    "accelerate>=1.6.0",
    "wandb>=0.25.1",
    "numpy>=2.4.0",
    "matplotlib>=3.10.0",
    "python-dotenv>=1.2.0",
    "pyyaml>=6.0.3",
    "pytest>=9.0.0",
    "ipykernel>=7.2.0",
]
